REFACTOR: 랭킹 관련 메소드 이동 (UserService -> MatrixService)
