+++ src/main/java/com/dnd/ground/domain/user/controller/UserControllerImpl.java
@@ -1,5 +1,6 @@
 package com.dnd.ground.domain.user.controller;
 
+import com.dnd.ground.domain.matrix.matrixService.MatrixService;
 import com.dnd.ground.domain.user.dto.HomeResponseDto;
 import com.dnd.ground.domain.user.dto.RankResponseDto;
 import com.dnd.ground.domain.user.service.UserService;
@@ -14,10 +15,11 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 /**
- * @description 메인홈 구성 컨트롤러 클래스
- * @author  박세헌
+ * @description 회원 관련 컨트롤러 구현체
+ * @author  박세헌, 박찬호
  * @since   2022-08-02
- * @updated 2022-08-09 / 랭킹 조회 api 추가: 박세헌
+ * @updated 1. 랭킹 관련 메소드 이동(UserService -> MatrixService)
+ *          - 2022.08.11 박찬호
  */
 
 @Api(tags = "유저")
@@ -36,18 +38,6 @@ public class UserControllerImpl implements UserController {
                     "나와 챌린지를 하는 유저 정보(challengeMatrices, 리스트) 조회")
     public ResponseEntity<HomeResponseDto> home(@RequestParam("nickname") String nickName){
         return ResponseEntity.ok(userService.showHome(nickName));
-    }
-
-    @Operation(summary = "칸의 수 랭킹", description = "칸의 수가 높은 순서대로 유저들을 조회")
-    @GetMapping("/rank/matrix")
-    public ResponseEntity<RankResponseDto.matrixRankingResponseDto> matrixRank(@RequestParam("nickname") String nickName){
-        return ResponseEntity.ok(userService.matrixRanking(nickName));
-    }
-
-    @Operation(summary = "영역의 수 랭킹", description = "영역의 수가 높은 순서대로 유저들을 조회")
-    @GetMapping("/rank/area")
-    public ResponseEntity<RankResponseDto.areaRankingResponseDto> areaRank(@RequestParam("nickname") String nickName){
-        return ResponseEntity.ok(userService.areaRanking(nickName));
     }
 
 }
