+++ src/main/java/com/dnd/ground/domain/user/service/UserService.java
@@ -5,15 +5,14 @@ import com.dnd.ground.domain.user.dto.HomeResponseDto;
 import com.dnd.ground.domain.user.dto.RankResponseDto;
 
 /**
- * @description 유저 서비스 인터페이스
+ * @description 회원 서비스 인터페이스
  * @author  박세헌, 박찬호
  * @since   2022-08-01
- * @updated 2022-08-09 / 랭킹 관련 함수 추가: 박세헌
+ * @updated 1. 랭킹 관련 메소드 이동(UserService -> MatrixService)
+ *          - 2022.08.11 박찬호
  */
 
 public interface UserService {
     User save(User user);
     HomeResponseDto showHome(String nickname);
-    RankResponseDto.matrixRankingResponseDto matrixRanking(String nickname);
-    RankResponseDto.areaRankingResponseDto areaRanking(String nickname);
 }
